Split virtual-method tests into minimal/full codegen files
