+++ itest/rust/src/object_tests/mod.rs
@@ -27,6 +27,7 @@ mod singleton_test;
 // `validate_property` is only supported in Godot 4.2+.
 #[cfg(since_api = "4.2")]
 mod validate_property_test;
+mod virtual_methods_niche_test;
 mod virtual_methods_test;
 
 // Need to test this in the init level method.
